Refactor print results engine
